requirements.txt
# assume all deps use semver, tighten otherwise
# for 0.x libs, assume that 0.k+1 breaks 0.k

toml>=0.10.1,<0.11.0

# for converting Telegram animated stickers to animated PNG
lottie>=0.6.6,<0.7.0
apng>=0.3.4,<0.4.0

# telegram bot deps
telethon>=1.16.4,<2.0.0

# signal deps
sqlcipher3-binary>=0.4.2,<0.5.0
# installing from git because my AnyIO PR has not been merged yet
signalstickers-client @ git+https://github.com/iomintz/signalstickers-client.git@pr/anyio
# installing from git because my AnyIO PR has not been pushed to PyPI yet, and my ctx.bot attribute PR
# has not been merged yet
semaphore-bot @ git+https://github.com/iomintz/semaphore.git@pr/add-ctx.bot
